респределил функции по файлам
