+++ src/nextMonthAndPrevMonth.ts
@@ -1,12 +1,12 @@
-import { RunTask } from "./api";
-import { renderCalendar } from "./index";
+import { currentDateHolder } from "./api";
+import { renderCalendar } from "./renderCalendar";
 
 export function nextMonth() {
   document.getElementById("nextBtn")?.addEventListener("click", () => {
-    RunTask.month++;
-    if (RunTask.month > 11) {
-      RunTask.month = 0;
-      RunTask.year++;
+    currentDateHolder.month++;
+    if (currentDateHolder.month > 11) {
+      currentDateHolder.month = 0;
+      currentDateHolder.year++;
     }
     renderCalendar();
   });
@@ -14,10 +14,10 @@ export function nextMonth() {
 
 export function prevMonth() {
   document.getElementById("prevBtn")?.addEventListener("click", () => {
-    RunTask.month--;
-    if (RunTask.month < 0) {
-      RunTask.month = 11;
-      RunTask.year--;
+    currentDateHolder.month--;
+    if (currentDateHolder.month < 0) {
+      currentDateHolder.month = 11;
+      currentDateHolder.year--;
     }
     renderCalendar();
   });
