+++ src/saveBtnTask.ts
@@ -1,6 +1,6 @@
 import { RunTask } from "./api";
-import { Task } from "./LSCalendarApi";
-import { selected, dateClicked } from "./currentClickDiv";
+import { type Task } from "./api";
+import { dateClicked } from "./currentClickDiv";
 
 //комит
 export async function saveBtnTask() {
@@ -18,10 +18,6 @@ export async function saveBtnTask() {
     const descriptionText = inputTextDescription.value.trim();
     e.preventDefault();
     if (taskText) {
-      const taskTitle = document.createElement("div");
-      taskTitle.className = "taskInCells";
-      selected.appendChild(taskTitle);
-      taskTitle.textContent = taskText;
       popupForm.style.display = "none";
       backgroundOverlay.style.display = "none";
       const secureID = crypto.randomUUID();
@@ -41,7 +37,7 @@ export async function saveBtnTask() {
   });
 }
 
-function addTask(title: string, description: string, date: string) {
+function addTaskToList(title: string, description: string, date: string) {
   const boxTask = document.getElementById("box-task");
   const itemTask = document.createElement("div"); //коробка для данных
   itemTask.className = "item-task";
@@ -62,10 +58,28 @@ function addTask(title: string, description: string, date: string) {
   dateTask.innerText = date;
   itemTask.appendChild(dateTask);
 }
+
+function addTaskToCalendar(title: string, description: string, date: string) {
+  const selected = document.querySelector(`[data-date="${date}"]`);
+  if (selected) {
+    const taskTitle = document.createElement("div");
+    taskTitle.className = "taskInCells";
+    selected.appendChild(taskTitle);
+    taskTitle.textContent = title;
+  }
+}
+
+function addTask(title: string, description: string, date: string) {
+  addTaskToList(title, description, date);
+
+  addTaskToCalendar(title, description, date);
+}
 function loadTasks() {
-  RunTask.getTasks().forEach((task: Task) => {
-    addTask(task.title, task.description, task.date);
-  });
+  RunTask.getTasks().then((tasks) =>
+    tasks.forEach((task: Task) => {
+      addTask(task.title, task.description, task.date);
+    }),
+  );
 }
 window.onload = loadTasks;
