--- src/updateMonthYearDisplay.ts
@@ -1,10 +1,0 @@
-import { RunTask } from "./api";
-import { currentMonth } from "./index";
-
-export async function updateMonthYearDisplay() {
-  const renderMonthYear = new Date(RunTask.year, RunTask.month);
-  currentMonth.textContent = renderMonthYear.toLocaleDateString("ru-RU", {
-    month: "long",
-    year: "numeric",
-  });
-}
